My take on the Inventory system (read desc for more detail)
I think my scriptableObject would not be as easy expendable as the last, but for the long run it would help the programmer easier add and remove, or even recognize what attribute is currently in use.
Also, my take on the inventory UI button and panel with working visual animate via code, looks pretty cool.